Application/requirements.txt
azure-common==1.1.28
azure-core==1.24.0
azure-functions==1.7.2
azure-identity==1.6.1
azure-keyvault-secrets==4.3.0
azure-mgmt-core==1.3.0
azure-mgmt-resource==21.0.0
azure-storage-blob==12.9.0b1
azure-storage-file-datalake==12.4.0
protobuf
six
azure-devops==5.1.0b1
pandas==1.1.5
numpy==1.19.5
pyarrow==6.0.0
requests==2.28.1
simplejson==3.19.1
tabula-py==2.7.0
pypdf2==3.0.1
bs4==0.0.1
datetime
tabulate==0.9.0
unidecode==1.3.6
azure-mgmt-datafactory==2.0.0